+++ src/refactoring/components/cart/index.ts
@@ -2,3 +2,4 @@ export * from "./CartItemCard";
 export * from "./CouponSelection";
 export * from "./OrderSummary";
 export * from "./ProductCard";
+export * from "./ProductCardList";
